GHC.Prim

Description

GHC's primitive types and operations. Use GHC.Exts from the base package instead of importing this module directly.

The word size story.

Haskell98 specifies that signed integers (type Int) must contain at least 30 bits. GHC always implements Int using the primitive type Int#, whose size equals the MachDeps.h constant WORD_SIZE_IN_BITS. This is normally set based on the config.h parameter SIZEOF_HSWORD, i.e., 32 bits on 32-bit machines, 64 bits on 64-bit machines. However, it can also be explicitly set to a smaller number, e.g., 31 bits, to allow the possibility of using tag bits. Currently GHC itself has only 32-bit and 64-bit variants, but 30 or 31-bit code can be exported as an external core file for use in other back ends.

GHC also implements a primitive unsigned integer type Word# which always has the same number of bits as Int#.

In addition, GHC supports families of explicit-sized integers and words at 8, 16, 32, and 64 bits, with the usual arithmetic operations, comparisons, and a range of conversions. The 8-bit and 16-bit sizes are always represented as Int# and Word#, and the operations implemented in terms of the the primops on these types, with suitable range restrictions on the results (using the narrow$n$Int# and narrow$n$Word# families of primops. The 32-bit sizes are represented using Int# and Word# when WORD_SIZE_IN_BITS $geq$ 32; otherwise, these are represented using distinct primitive types Int32# and Word32#. These (when needed) have a complete set of corresponding operations; however, nearly all of these are implemented as external C functions rather than as primops. Exactly the same story applies to the 64-bit sizes. All of these details are hidden under the PrelInt and PrelWord modules, which use #if-defs to invoke the appropriate types and operators.

Word size also matters for the families of primops for indexing/reading/writing fixed-size quantities at offsets from an array base, address, or foreign pointer. Here, a slightly different approach is taken. The names of these primops are fixed, but their types vary according to the value of WORD_SIZE_IN_BITS. For example, if word size is at least 32 bits then an operator like indexInt32Array# has type ByteArray# -> Int# -> Int#; otherwise it has type ByteArray# -> Int# -> Int32#. This approach confines the necessary #if-defs to this file; no conditional compilation is needed in the files that expose these primops.

Finally, there are strongly deprecated primops for coercing between Addr#, the primitive type of machine addresses, and Int#. These are pretty bogus anyway, but will work on existing 32-bit and 64-bit GHC targets; they are completely bogus when tag bits are used in Int#, so are not available in this case.

Char#

Operations on 31-bit characters.

data Char#

gtChar# :: Char# -> Char# -> Bool

geChar# :: Char# -> Char# -> Bool

eqChar# :: Char# -> Char# -> Bool

neChar# :: Char# -> Char# -> Bool

ltChar# :: Char# -> Char# -> Bool

leChar# :: Char# -> Char# -> Bool

ord# :: Char# -> Int#

Int#

Operations on native-size integers (30+ bits).

data Int#

(+#) :: Int# -> Int# -> Int#

(-#) :: Int# -> Int# -> Int#

(*#) :: Int# -> Int# -> Int#

Low word of signed integer multiply.

mulIntMayOflo# :: Int# -> Int# -> Int#

Return non-zero if there is any possibility that the upper word of a signed integer multiply might contain useful information. Return zero only if you are completely sure that no overflow can occur. On a 32-bit platform, the recommmended implementation is to do a 32 x 32 -> 64 signed multiply, and subtract result[63:32] from (result[31] >>signed 31). If this is zero, meaning that the upper word is merely a sign extension of the lower one, no overflow can occur.

On a 64-bit platform it is not always possible to acquire the top 64 bits of the result. Therefore, a recommended implementation is to take the absolute value of both operands, and return 0 iff bits[63:31] of them are zero, since that means that their magnitudes fit within 31 bits, so the magnitude of the product must fit into 62 bits.

If in doubt, return non-zero, but do make an effort to create the correct answer for small args, since otherwise the performance of (*) :: Integer -> Integer -> Integer will be poor.

quotInt# :: Int# -> Int# -> Int#

Rounds towards zero.

remInt# :: Int# -> Int# -> Int#

Satisfies (quotInt# x y) *# y +# (remInt# x y) == x.

quotRemInt# :: Int# -> Int# -> (#Int#, Int##)

Rounds towards zero.

negateInt# :: Int# -> Int#

addIntC# :: Int# -> Int# -> (#Int#, Int##)

Add with carry. First member of result is (wrapped) sum; second member is 0 iff no overflow occured.

subIntC# :: Int# -> Int# -> (#Int#, Int##)

Subtract with carry. First member of result is (wrapped) difference; second member is 0 iff no overflow occured.

(>#) :: Int# -> Int# -> Bool

(>=#) :: Int# -> Int# -> Bool

(==#) :: Int# -> Int# -> Bool

(/=#) :: Int# -> Int# -> Bool

(<#) :: Int# -> Int# -> Bool

(<=#) :: Int# -> Int# -> Bool

chr# :: Int# -> Char#

int2Word# :: Int# -> Word#

int2Float# :: Int# -> Float#

int2Double# :: Int# -> Double#

uncheckedIShiftL# :: Int# -> Int# -> Int#

Shift left. Result undefined if shift amount is not in the range 0 to word size - 1 inclusive.

uncheckedIShiftRA# :: Int# -> Int# -> Int#

Shift right arithmetic. Result undefined if shift amount is not in the range 0 to word size - 1 inclusive.

uncheckedIShiftRL# :: Int# -> Int# -> Int#

Shift right logical. Result undefined if shift amount is not in the range 0 to word size - 1 inclusive.

Word#

Operations on native-sized unsigned words (30+ bits).

data Word#

plusWord# :: Word# -> Word# -> Word#

plusWord2# :: Word# -> Word# -> (#Word#, Word##)

minusWord# :: Word# -> Word# -> Word#

timesWord# :: Word# -> Word# -> Word#

timesWord2# :: Word# -> Word# -> (#Word#, Word##)

quotWord# :: Word# -> Word# -> Word#

remWord# :: Word# -> Word# -> Word#

quotRemWord# :: Word# -> Word# -> (#Word#, Word##)

quotRemWord2# :: Word# -> Word# -> Word# -> (#Word#, Word##)

and# :: Word# -> Word# -> Word#

or# :: Word# -> Word# -> Word#

xor# :: Word# -> Word# -> Word#

not# :: Word# -> Word#

uncheckedShiftL# :: Word# -> Int# -> Word#

Shift left logical. Result undefined if shift amount is not in the range 0 to word size - 1 inclusive.

uncheckedShiftRL# :: Word# -> Int# -> Word#

Shift right logical. Result undefined if shift amount is not in the range 0 to word size - 1 inclusive.

word2Int# :: Word# -> Int#

gtWord# :: Word# -> Word# -> Bool

geWord# :: Word# -> Word# -> Bool

eqWord# :: Word# -> Word# -> Bool

neWord# :: Word# -> Word# -> Bool

ltWord# :: Word# -> Word# -> Bool

leWord# :: Word# -> Word# -> Bool

popCnt8# :: Word# -> Word#

Count the number of set bits in the lower 8 bits of a word.

popCnt16# :: Word# -> Word#

Count the number of set bits in the lower 16 bits of a word.

popCnt32# :: Word# -> Word#

Count the number of set bits in the lower 32 bits of a word.

popCnt64# :: Word# -> Word#

Count the number of set bits in a 64-bit word.

popCnt# :: Word# -> Word#

Count the number of set bits in a word.

Narrowings

Explicit narrowing of native-sized ints or words.

narrow8Int# :: Int# -> Int#

narrow16Int# :: Int# -> Int#

narrow32Int# :: Int# -> Int#

narrow8Word# :: Word# -> Word#

narrow16Word# :: Word# -> Word#

narrow32Word# :: Word# -> Word#

Double#

Operations on double-precision (64 bit) floating-point numbers.

data Double#

(>##) :: Double# -> Double# -> Bool

(>=##) :: Double# -> Double# -> Bool

(==##) :: Double# -> Double# -> Bool

(/=##) :: Double# -> Double# -> Bool

(<##) :: Double# -> Double# -> Bool

(<=##) :: Double# -> Double# -> Bool

(+##) :: Double# -> Double# -> Double#

(-##) :: Double# -> Double# -> Double#

(*##) :: Double# -> Double# -> Double#

(/##) :: Double# -> Double# -> Double#

negateDouble# :: Double# -> Double#

double2Int# :: Double# -> Int#

Truncates a Double. Results are undefined if the truncation if truncation yields a value outside the range of Int#.

double2Float# :: Double# -> Float#

expDouble# :: Double# -> Double#

logDouble# :: Double# -> Double#

sqrtDouble# :: Double# -> Double#

sinDouble# :: Double# -> Double#

cosDouble# :: Double# -> Double#

tanDouble# :: Double# -> Double#

asinDouble# :: Double# -> Double#

acosDouble# :: Double# -> Double#

atanDouble# :: Double# -> Double#

sinhDouble# :: Double# -> Double#

coshDouble# :: Double# -> Double#

tanhDouble# :: Double# -> Double#

(**##) :: Double# -> Double# -> Double#

Exponentiation.

decodeDouble_2Int# :: Double# -> (#Int#, Word#, Word#, Int##)

Convert to integer. First component of the result is -1 or 1, indicating the sign of the mantissa. The next two are the high and low 32 bits of the mantissa respectively, and the last is the exponent.

Float#

Operations on single-precision (32-bit) floating-point numbers.

data Float#

gtFloat# :: Float# -> Float# -> Bool

geFloat# :: Float# -> Float# -> Bool

eqFloat# :: Float# -> Float# -> Bool

neFloat# :: Float# -> Float# -> Bool

ltFloat# :: Float# -> Float# -> Bool

leFloat# :: Float# -> Float# -> Bool

plusFloat# :: Float# -> Float# -> Float#

minusFloat# :: Float# -> Float# -> Float#

timesFloat# :: Float# -> Float# -> Float#

divideFloat# :: Float# -> Float# -> Float#

negateFloat# :: Float# -> Float#

float2Int# :: Float# -> Int#

Truncates a Float. Results are undefined if the truncation if truncation yields a value outside the range of Int#.

expFloat# :: Float# -> Float#

logFloat# :: Float# -> Float#

sqrtFloat# :: Float# -> Float#

sinFloat# :: Float# -> Float#

cosFloat# :: Float# -> Float#

tanFloat# :: Float# -> Float#

asinFloat# :: Float# -> Float#

acosFloat# :: Float# -> Float#

atanFloat# :: Float# -> Float#

sinhFloat# :: Float# -> Float#

coshFloat# :: Float# -> Float#

tanhFloat# :: Float# -> Float#

powerFloat# :: Float# -> Float# -> Float#

float2Double# :: Float# -> Double#

decodeFloat_Int# :: Float# -> (#Int#, Int##)

Convert to integers. First Int# in result is the mantissa; second is the exponent.

Arrays

Operations on Array#.

data Array# a

data MutableArray# s a

newArray# :: Int# -> a -> State# s -> (#State# s, MutableArray# s a#)

Create a new mutable array with the specified number of elements, in the specified state thread, with each element containing the specified initial value.

sameMutableArray# :: MutableArray# s a -> MutableArray# s a -> Bool

readArray# :: MutableArray# s a -> Int# -> State# s -> (#State# s, a#)

Read from specified index of mutable array. Result is not yet evaluated.

writeArray# :: MutableArray# s a -> Int# -> a -> State# s -> State# s

Write to specified index of mutable array.

sizeofArray# :: Array# a -> Int#

Return the number of elements in the array.

sizeofMutableArray# :: MutableArray# s a -> Int#

Return the number of elements in the array.

indexArray# :: Array# a -> Int# -> (#a#)

Read from specified index of immutable array. Result is packaged into an unboxed singleton; the result itself is not yet evaluated.

unsafeFreezeArray# :: MutableArray# s a -> State# s -> (#State# s, Array# a#)

Make a mutable array immutable, without copying.

unsafeThawArray# :: Array# a -> State# s -> (#State# s, MutableArray# s a#)

Make an immutable array mutable, without copying.

copyArray# :: Array# a -> Int# -> MutableArray# s a -> Int# -> Int# -> State# s -> State# s

Copy a range of the Array. Both arrays must fully contain the specified ranges, but this is not checked. The two arrays must not be the same array in different states, but this is not checked either.

copyMutableArray# :: MutableArray# s a -> Int# -> MutableArray# s a -> Int# -> Int# -> State# s -> State# s

Copy a range of the first MutableArray. Both arrays must fully contain the specified ranges, but this is not checked.

cloneArray# :: Array# a -> Int# -> Int# -> Array# a

Return a newly allocated Array. The provided Arrays, but this is not checked.

cloneMutableArray# :: MutableArray# s a -> Int# -> Int# -> State# s -> (#State# s, MutableArray# s a#)

Return a newly allocated Array. The provided MutableArrays, but this is not checked.

freezeArray# :: MutableArray# s a -> Int# -> Int# -> State# s -> (#State# s, Array# a#)

Return a newly allocated Array. The provided MutableArrays, but this is not checked.

thawArray# :: Array# a -> Int# -> Int# -> State# s -> (#State# s, MutableArray# s a#)

Return a newly allocated Array. The provided Arrays, but this is not checked.

Byte Arrays

Operations on ByteArray#. A ByteArray# is a just a region of raw memory in the garbage-collected heap, which is not scanned for pointers. It carries its own size (in bytes). There are three sets of operations for accessing byte array contents: index for reading from immutable byte arrays, and read/write for mutable byte arrays. Each set contains operations for a range of useful primitive data types. Each operation takes an offset measured in terms of the size of the primitive type being read or written.

data ByteArray#

data MutableByteArray# s

newByteArray# :: Int# -> State# s -> (#State# s, MutableByteArray# s#)

Create a new mutable byte array of specified size (in bytes), in the specified state thread.

newPinnedByteArray# :: Int# -> State# s -> (#State# s, MutableByteArray# s#)

Create a mutable byte array that the GC guarantees not to move.

newAlignedPinnedByteArray# :: Int# -> Int# -> State# s -> (#State# s, MutableByteArray# s#)

Create a mutable byte array, aligned by the specified amount, that the GC guarantees not to move.

byteArrayContents# :: ByteArray# -> Addr#

Intended for use with pinned arrays; otherwise very unsafe!

sameMutableByteArray# :: MutableByteArray# s -> MutableByteArray# s -> Bool

unsafeFreezeByteArray# :: MutableByteArray# s -> State# s -> (#State# s, ByteArray##)

Make a mutable byte array immutable, without copying.

sizeofByteArray# :: ByteArray# -> Int#

Return the size of the array in bytes.

sizeofMutableByteArray# :: MutableByteArray# s -> Int#

Return the size of the array in bytes.

indexCharArray# :: ByteArray# -> Int# -> Char#

Read 8-bit character; offset in bytes.

indexWideCharArray# :: ByteArray# -> Int# -> Char#

Read 31-bit character; offset in 4-byte words.

indexIntArray# :: ByteArray# -> Int# -> Int#

indexWordArray# :: ByteArray# -> Int# -> Word#

indexAddrArray# :: ByteArray# -> Int# -> Addr#

indexFloatArray# :: ByteArray# -> Int# -> Float#

indexDoubleArray# :: ByteArray# -> Int# -> Double#

indexStablePtrArray# :: ByteArray# -> Int# -> StablePtr# a

indexInt8Array# :: ByteArray# -> Int# -> Int#

indexInt16Array# :: ByteArray# -> Int# -> Int#

indexInt32Array# :: ByteArray# -> Int# -> Int#

indexInt64Array# :: ByteArray# -> Int# -> Int#

indexWord8Array# :: ByteArray# -> Int# -> Word#

indexWord16Array# :: ByteArray# -> Int# -> Word#

indexWord32Array# :: ByteArray# -> Int# -> Word#

indexWord64Array# :: ByteArray# -> Int# -> Word#

readCharArray# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, Char##)

Read 8-bit character; offset in bytes.

readWideCharArray# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, Char##)

Read 31-bit character; offset in 4-byte words.

readIntArray# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, Int##)

readWordArray# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, Word##)

readAddrArray# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, Addr##)

readFloatArray# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, Float##)

readDoubleArray# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, Double##)

readStablePtrArray# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, StablePtr# a#)

readInt8Array# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, Int##)

readInt16Array# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, Int##)

readInt32Array# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, Int##)

readInt64Array# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, Int##)

readWord8Array# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, Word##)

readWord16Array# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, Word##)

readWord32Array# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, Word##)

readWord64Array# :: MutableByteArray# s -> Int# -> State# s -> (#State# s, Word##)

writeCharArray# :: MutableByteArray# s -> Int# -> Char# -> State# s -> State# s

Write 8-bit character; offset in bytes.

writeWideCharArray# :: MutableByteArray# s -> Int# -> Char# -> State# s -> State# s

Write 31-bit character; offset in 4-byte words.

writeIntArray# :: MutableByteArray# s -> Int# -> Int# -> State# s -> State# s

writeWordArray# :: MutableByteArray# s -> Int# -> Word# -> State# s -> State# s

writeAddrArray# :: MutableByteArray# s -> Int# -> Addr# -> State# s -> State# s

writeFloatArray# :: MutableByteArray# s -> Int# -> Float# -> State# s -> State# s

writeDoubleArray# :: MutableByteArray# s -> Int# -> Double# -> State# s -> State# s

writeStablePtrArray# :: MutableByteArray# s -> Int# -> StablePtr# a -> State# s -> State# s

writeInt8Array# :: MutableByteArray# s -> Int# -> Int# -> State# s -> State# s

writeInt16Array# :: MutableByteArray# s -> Int# -> Int# -> State# s -> State# s

writeInt32Array# :: MutableByteArray# s -> Int# -> Int# -> State# s -> State# s

writeInt64Array# :: MutableByteArray# s -> Int# -> Int# -> State# s -> State# s

writeWord8Array# :: MutableByteArray# s -> Int# -> Word# -> State# s -> State# s

writeWord16Array# :: MutableByteArray# s -> Int# -> Word# -> State# s -> State# s

writeWord32Array# :: MutableByteArray# s -> Int# -> Word# -> State# s -> State# s

writeWord64Array# :: MutableByteArray# s -> Int# -> Word# -> State# s -> State# s

copyByteArray# :: ByteArray# -> Int# -> MutableByteArray# s -> Int# -> Int# -> State# s -> State# s

Copy a range of the ByteArray. Both arrays must fully contain the specified ranges, but this is not checked. The two arrays must not be the same array in different states, but this is not checked either.

copyMutableByteArray# :: MutableByteArray# s -> Int# -> MutableByteArray# s -> Int# -> Int# -> State# s -> State# s

Copy a range of the first MutableByteArray. Both arrays must fully contain the specified ranges, but this is not checked.

setByteArray# :: MutableByteArray# s -> Int# -> Int# -> Int# -> State# s -> State# s

Set the range of the MutableByteArray# to the specified character.

Arrays of arrays

Operations on ArrayArray#. An ArrayArray# contains references to {em unpointed} arrays, such as ByteArray#s. Hence, it is not parameterised by the element types, just like a ByteArray#, but it needs to be scanned during GC, just like an Array#. We represent an ArrayArray# exactly as a Array#, but provide element-type-specific indexing, reading, and writing.

data ArrayArray#

data MutableArrayArray# s

newArrayArray# :: Int# -> State# s -> (#State# s, MutableArrayArray# s#)

Create a new mutable array of arrays with the specified number of elements, in the specified state thread, with each element recursively referring to the newly created array.

sameMutableArrayArray# :: MutableArrayArray# s -> MutableArrayArray# s -> Bool

unsafeFreezeArrayArray# :: MutableArrayArray# s -> State# s -> (#State# s, ArrayArray##)

Make a mutable array of arrays immutable, without copying.

sizeofArrayArray# :: ArrayArray# -> Int#

Return the number of elements in the array.

sizeofMutableArrayArray# :: MutableArrayArray# s -> Int#

Return the number of elements in the array.

indexByteArrayArray# :: ArrayArray# -> Int# -> ByteArray#

indexArrayArrayArray# :: ArrayArray# -> Int# -> ArrayArray#

readByteArrayArray# :: MutableArrayArray# s -> Int# -> State# s -> (#State# s, ByteArray##)

readMutableByteArrayArray# :: MutableArrayArray# s -> Int# -> State# s -> (#State# s, MutableByteArray# s#)

readArrayArrayArray# :: MutableArrayArray# s -> Int# -> State# s -> (#State# s, ArrayArray##)

readMutableArrayArrayArray# :: MutableArrayArray# s -> Int# -> State# s -> (#State# s, MutableArrayArray# s#)

writeByteArrayArray# :: MutableArrayArray# s -> Int# -> ByteArray# -> State# s -> State# s

writeMutableByteArrayArray# :: MutableArrayArray# s -> Int# -> MutableByteArray# s -> State# s -> State# s

writeArrayArrayArray# :: MutableArrayArray# s -> Int# -> ArrayArray# -> State# s -> State# s

writeMutableArrayArrayArray# :: MutableArrayArray# s -> Int# -> MutableArrayArray# s -> State# s -> State# s

copyArrayArray# :: ArrayArray# -> Int# -> MutableArrayArray# s -> Int# -> Int# -> State# s -> State# s

Copy a range of the ArrayArray. Both arrays must fully contain the specified ranges, but this is not checked. The two arrays must not be the same array in different states, but this is not checked either.

copyMutableArrayArray# :: MutableArrayArray# s -> Int# -> MutableArrayArray# s -> Int# -> Int# -> State# s -> State# s

Copy a range of the first MutableArrayArray# to the specified region in the second MutableArrayArray#. Both arrays must fully contain the specified ranges, but this is not checked.

Addr#

data Addr#

An arbitrary machine address assumed to point outside the garbage-collected heap.

nullAddr# :: Addr#

The null address.

plusAddr# :: Addr# -> Int# -> Addr#

minusAddr# :: Addr# -> Addr# -> Int#

Result is meaningless if two Addr#s are so far apart that their difference doesn't fit in an Int#.

remAddr# :: Addr# -> Int# -> Int#

Return the remainder when the Addr# arg, treated like an Int#, is divided by the Int# arg.

addr2Int# :: Addr# -> Int#

Coerce directly from address to int. Strongly deprecated.

int2Addr# :: Int# -> Addr#

Coerce directly from int to address. Strongly deprecated.

gtAddr# :: Addr# -> Addr# -> Bool

geAddr# :: Addr# -> Addr# -> Bool

eqAddr# :: Addr# -> Addr# -> Bool

neAddr# :: Addr# -> Addr# -> Bool

ltAddr# :: Addr# -> Addr# -> Bool

leAddr# :: Addr# -> Addr# -> Bool

indexCharOffAddr# :: Addr# -> Int# -> Char#

Reads 8-bit character; offset in bytes.

indexWideCharOffAddr# :: Addr# -> Int# -> Char#

Reads 31-bit character; offset in 4-byte words.

indexIntOffAddr# :: Addr# -> Int# -> Int#

indexWordOffAddr# :: Addr# -> Int# -> Word#

indexAddrOffAddr# :: Addr# -> Int# -> Addr#

indexFloatOffAddr# :: Addr# -> Int# -> Float#

indexDoubleOffAddr# :: Addr# -> Int# -> Double#

indexStablePtrOffAddr# :: Addr# -> Int# -> StablePtr# a

indexInt8OffAddr# :: Addr# -> Int# -> Int#

indexInt16OffAddr# :: Addr# -> Int# -> Int#

indexInt32OffAddr# :: Addr# -> Int# -> Int#

indexInt64OffAddr# :: Addr# -> Int# -> Int#

indexWord8OffAddr# :: Addr# -> Int# -> Word#

indexWord16OffAddr# :: Addr# -> Int# -> Word#

indexWord32OffAddr# :: Addr# -> Int# -> Word#

indexWord64OffAddr# :: Addr# -> Int# -> Word#

readCharOffAddr# :: Addr# -> Int# -> State# s -> (#State# s, Char##)

Reads 8-bit character; offset in bytes.

readWideCharOffAddr# :: Addr# -> Int# -> State# s -> (#State# s, Char##)

Reads 31-bit character; offset in 4-byte words.

readIntOffAddr# :: Addr# -> Int# -> State# s -> (#State# s, Int##)

readWordOffAddr# :: Addr# -> Int# -> State# s -> (#State# s, Word##)

readAddrOffAddr# :: Addr# -> Int# -> State# s -> (#State# s, Addr##)

readFloatOffAddr# :: Addr# -> Int# -> State# s -> (#State# s, Float##)

readDoubleOffAddr# :: Addr# -> Int# -> State# s -> (#State# s, Double##)

readStablePtrOffAddr# :: Addr# -> Int# -> State# s -> (#State# s, StablePtr# a#)

readInt8OffAddr# :: Addr# -> Int# -> State# s -> (#State# s, Int##)

readInt16OffAddr# :: Addr# -> Int# -> State# s -> (#State# s, Int##)

readInt32OffAddr# :: Addr# -> Int# -> State# s -> (#State# s, Int##)

readInt64OffAddr# :: Addr# -> Int# -> State# s -> (#State# s, Int##)

readWord8OffAddr# :: Addr# -> Int# -> State# s -> (#State# s, Word##)

readWord16OffAddr# :: Addr# -> Int# -> State# s -> (#State# s, Word##)

readWord32OffAddr# :: Addr# -> Int# -> State# s -> (#State# s, Word##)

readWord64OffAddr# :: Addr# -> Int# -> State# s -> (#State# s, Word##)

writeCharOffAddr# :: Addr# -> Int# -> Char# -> State# s -> State# s

writeWideCharOffAddr# :: Addr# -> Int# -> Char# -> State# s -> State# s

writeIntOffAddr# :: Addr# -> Int# -> Int# -> State# s -> State# s

writeWordOffAddr# :: Addr# -> Int# -> Word# -> State# s -> State# s

writeAddrOffAddr# :: Addr# -> Int# -> Addr# -> State# s -> State# s

writeFloatOffAddr# :: Addr# -> Int# -> Float# -> State# s -> State# s

writeDoubleOffAddr# :: Addr# -> Int# -> Double# -> State# s -> State# s

writeStablePtrOffAddr# :: Addr# -> Int# -> StablePtr# a -> State# s -> State# s

writeInt8OffAddr# :: Addr# -> Int# -> Int# -> State# s -> State# s

writeInt16OffAddr# :: Addr# -> Int# -> Int# -> State# s -> State# s

writeInt32OffAddr# :: Addr# -> Int# -> Int# -> State# s -> State# s

writeInt64OffAddr# :: Addr# -> Int# -> Int# -> State# s -> State# s

writeWord8OffAddr# :: Addr# -> Int# -> Word# -> State# s -> State# s

writeWord16OffAddr# :: Addr# -> Int# -> Word# -> State# s -> State# s

writeWord32OffAddr# :: Addr# -> Int# -> Word# -> State# s -> State# s

writeWord64OffAddr# :: Addr# -> Int# -> Word# -> State# s -> State# s

Mutable variables

Operations on MutVar#s.

data MutVar# s a

A MutVar# behaves like a single-element mutable array.

newMutVar# :: a -> State# s -> (#State# s, MutVar# s a#)

Create MutVar# with specified initial value in specified state thread.

readMutVar# :: MutVar# s a -> State# s -> (#State# s, a#)

Read contents of MutVar#. Result is not yet evaluated.

writeMutVar# :: MutVar# s a -> a -> State# s -> State# s

Write contents of MutVar#.

sameMutVar# :: MutVar# s a -> MutVar# s a -> Bool

atomicModifyMutVar# :: MutVar# s a -> (a -> b) -> State# s -> (#State# s, c#)

casMutVar# :: MutVar# s a -> a -> a -> State# s -> (#State# s, Int#, a#)

Exceptions

catch# :: (State# RealWorld -> (#State# RealWorld, a#)) -> (b -> State# RealWorld -> (#State# RealWorld, a#)) -> State# RealWorld -> (#State# RealWorld, a#)

raise# :: a -> b

raiseIO# :: a -> State# RealWorld -> (#State# RealWorld, b#)

maskAsyncExceptions# :: (State# RealWorld -> (#State# RealWorld, a#)) -> State# RealWorld -> (#State# RealWorld, a#)

maskUninterruptible# :: (State# RealWorld -> (#State# RealWorld, a#)) -> State# RealWorld -> (#State# RealWorld, a#)

unmaskAsyncExceptions# :: (State# RealWorld -> (#State# RealWorld, a#)) -> State# RealWorld -> (#State# RealWorld, a#)

getMaskingState# :: State# RealWorld -> (#State# RealWorld, Int##)

STM-accessible Mutable Variables

data TVar# s a

atomically# :: (State# RealWorld -> (#State# RealWorld, a#)) -> State# RealWorld -> (#State# RealWorld, a#)

retry# :: State# RealWorld -> (#State# RealWorld, a#)

catchRetry# :: (State# RealWorld -> (#State# RealWorld, a#)) -> (State# RealWorld -> (#State# RealWorld, a#)) -> State# RealWorld -> (#State# RealWorld, a#)

catchSTM# :: (State# RealWorld -> (#State# RealWorld, a#)) -> (b -> State# RealWorld -> (#State# RealWorld, a#)) -> State# RealWorld -> (#State# RealWorld, a#)

check# :: (State# RealWorld -> (#State# RealWorld, a#)) -> State# RealWorld -> (#State# RealWorld, ()#)

newTVar# :: a -> State# s -> (#State# s, TVar# s a#)

Create a new TVar# holding a specified initial value.

readTVar# :: TVar# s a -> State# s -> (#State# s, a#)

Read contents of TVar#. Result is not yet evaluated.

readTVarIO# :: TVar# s a -> State# s -> (#State# s, a#)

Read contents of TVar# outside an STM transaction

writeTVar# :: TVar# s a -> a -> State# s -> State# s

Write contents of TVar#.

sameTVar# :: TVar# s a -> TVar# s a -> Bool

Synchronized Mutable Variables

Operations on MVar#s.

data MVar# s a

A shared mutable variable (not the same as a MutVar#!). (Note: in a non-concurrent implementation, (MVar# a) can be represented by (MutVar# (Maybe a)).)

newMVar# :: State# s -> (#State# s, MVar# s a#)

Create new MVar#; initially empty.

takeMVar# :: MVar# s a -> State# s -> (#State# s, a#)

If MVar# is empty, block until it becomes full. Then remove and return its contents, and set it empty.

tryTakeMVar# :: MVar# s a -> State# s -> (#State# s, Int#, a#)

If MVar# is empty, immediately return with integer 0 and value undefined. Otherwise, return with integer 1 and contents of MVar#, and set MVar# empty.

putMVar# :: MVar# s a -> a -> State# s -> State# s

If MVar# is full, block until it becomes empty. Then store value arg as its new contents.

tryPutMVar# :: MVar# s a -> a -> State# s -> (#State# s, Int##)

If MVar# is full, immediately return with integer 0. Otherwise, store value arg as MVar#'s new contents, and return with integer 1.

sameMVar# :: MVar# s a -> MVar# s a -> Bool

isEmptyMVar# :: MVar# s a -> State# s -> (#State# s, Int##)

Return 1 if MVar# is empty; 0 otherwise.

Delay/wait operations

delay# :: Int# -> State# s -> State# s

Sleep specified number of microseconds.

waitRead# :: Int# -> State# s -> State# s

Block until input is available on specified file descriptor.

waitWrite# :: Int# -> State# s -> State# s

Block until output is possible on specified file descriptor.

Concurrency primitives

data State# s

State# is the primitive, unlifted type of states. It has one type parameter, thus State# RealWorld, or State# s, where s is a type variable. The only purpose of the type parameter is to keep different state threads separate. It is represented by nothing at all.

data RealWorld

RealWorld is deeply magical. It is primitive, but it is not unlifted (hence ptrArg). We never manipulate values of type RealWorld; it's only used in the type system, to parameterise State#.

data ThreadId#

(In a non-concurrent implementation, this can be a singleton type, whose (unique) value is returned by myThreadId#. The other operations can be omitted.)

fork# :: a -> State# RealWorld -> (#State# RealWorld, ThreadId##)

forkOn# :: Int# -> a -> State# RealWorld -> (#State# RealWorld, ThreadId##)

killThread# :: ThreadId# -> a -> State# RealWorld -> State# RealWorld

yield# :: State# RealWorld -> State# RealWorld

myThreadId# :: State# RealWorld -> (#State# RealWorld, ThreadId##)

labelThread# :: ThreadId# -> Addr# -> State# RealWorld -> State# RealWorld

isCurrentThreadBound# :: State# RealWorld -> (#State# RealWorld, Int##)

noDuplicate# :: State# RealWorld -> State# RealWorld

threadStatus# :: ThreadId# -> State# RealWorld -> (#State# RealWorld, Int#, Int#, Int##)

Weak pointers

data Weak# b

mkWeak# :: o -> b -> c -> State# RealWorld -> (#State# RealWorld, Weak# b#)

mkWeakNoFinalizer# :: o -> b -> State# RealWorld -> (#State# RealWorld, Weak# b#)

mkWeakForeignEnv# :: o -> b -> Addr# -> Addr# -> Int# -> Addr# -> State# RealWorld -> (#State# RealWorld, Weak# b#)

deRefWeak# :: Weak# a -> State# RealWorld -> (#State# RealWorld, Int#, a#)

finalizeWeak# :: Weak# a -> State# RealWorld -> (#State# RealWorld, Int#, State# RealWorld -> (#State# RealWorld, ()#)#)

touch# :: o -> State# RealWorld -> State# RealWorld

Stable pointers and names

data StablePtr# a

data StableName# a

makeStablePtr# :: a -> State# RealWorld -> (#State# RealWorld, StablePtr# a#)

deRefStablePtr# :: StablePtr# a -> State# RealWorld -> (#State# RealWorld, a#)

eqStablePtr# :: StablePtr# a -> StablePtr# a -> Int#

makeStableName# :: a -> State# RealWorld -> (#State# RealWorld, StableName# a#)

eqStableName# :: StableName# a -> StableName# a -> Int#

stableNameToInt# :: StableName# a -> Int#

Unsafe pointer equality

reallyUnsafePtrEquality# :: a -> a -> Int#

Parallelism

par# :: a -> Int#

spark# :: a -> State# s -> (#State# s, a#)

seq# :: a -> State# s -> (#State# s, a#)

getSpark# :: State# s -> (#State# s, Int#, a#)

numSparks# :: State# s -> (#State# s, Int##)

Returns the number of sparks in the local spark pool.

parGlobal# :: a -> Int# -> Int# -> Int# -> Int# -> b -> Int#

parLocal# :: a -> Int# -> Int# -> Int# -> Int# -> b -> Int#

parAt# :: b -> a -> Int# -> Int# -> Int# -> Int# -> c -> Int#

parAtAbs# :: a -> Int# -> Int# -> Int# -> Int# -> Int# -> b -> Int#

parAtRel# :: a -> Int# -> Int# -> Int# -> Int# -> Int# -> b -> Int#

parAtForNow# :: b -> a -> Int# -> Int# -> Int# -> Int# -> c -> Int#

Tag to enum stuff

Convert back and forth between values of enumerated types and small integers.

dataToTag# :: a -> Int#

tagToEnum# :: Int# -> a

Bytecode operations

Support for the bytecode interpreter and linker.

data BCO#

Primitive bytecode type.

addrToAny# :: Addr# -> (#a#)

Convert an Addr# to a followable Any type.

mkApUpd0# :: BCO# -> (#a#)

newBCO# :: ByteArray# -> ByteArray# -> Array# a -> Int# -> ByteArray# -> State# s -> (#State# s, BCO##)

unpackClosure# :: a -> (#Addr#, Array# b, ByteArray##)

getApStackVal# :: a -> Int# -> (#Int#, b#)

Misc

These aren't nearly as wired in as Etc...

getCCSOf# :: a -> State# s -> (#State# s, Addr##)

getCurrentCCS# :: a -> State# s -> (#State# s, Addr##)

Returns the current CostCentreStack (value is NULL if not profiling). Takes a dummy argument which can be used to avoid the call to getCCCS# being floated out by the simplifier, which would result in an uninformative stack ("CAF").

Etc

Miscellaneous built-ins

seq :: a -> b -> b

Evaluates its first argument to head normal form, and then returns its second argument as the result.

inline :: a -> a

The call (inline f) arranges that f is inlined, regardless of its size. More precisely, the call (inline f) rewrites to the right-hand side of f's definition. This allows the programmer to control inlining from a particular call site rather than the definition site of the function (c.f. INLINE pragmas in User's Guide, Section 7.10.3, "INLINE and NOINLINE pragmas").

This inlining occurs regardless of the argument to the call or the size of f's definition; it is unconditional. The main caveat is that f's definition must be visible to the compiler. That is, f must be let-bound in the current scope. If no inlining takes place, the inline function expands to the identity function in Phase zero; so its use imposes no overhead.

It is good practice to mark the function with an INLINABLE pragma at its definition, (a) so that GHC guarantees to expose its unfolding regardless of size, and (b) so that you have control over exactly what is inlined.

lazy :: a -> a

The lazy function restrains strictness analysis a little. The call (lazy e) means the same as e, but lazy has a magical property so far as strictness analysis is concerned: it is lazy in its first argument, even though its semantics is strict. After strictness analysis has run, calls to lazy are inlined to be the identity function.

This behaviour is occasionally useful when controlling evaluation order. Notably, lazy is used in the library definition of Control.Parallel.par:

par :: a -> b -> b

par x y = case (par# x) of _ -> lazy y

If lazy were not lazy, par would look strict in y which would defeat the whole purpose of par.

Like seq, the argument of lazy can have an unboxed type.

data Any k

The type constructor Any is type to which you can unsafely coerce any lifted type, and back.

It is lifted, and hence represented by a pointer

It does not claim to be a data type, and that's important for the code generator, because the code gen may enter a data value but never enters a function value.

It's also used to instantiate un-constrained type variables after type checking. For example, length has type

length :: forall a. [a] -> Int

and the list datacon for the empty list has type

[] :: forall a. [a]

In order to compose these two terms as length [] a type application is required, but there is no constraint on the choice. In this situation GHC uses Any:

length (Any *) ([] (Any *))

Note that Any is kind polymorphic, and takes a kind k as its first argument. The kind of Any is thus forall k. k -> k.

data AnyK

The kind AnyK is the kind level counterpart to Any. In a kind polymorphic setting, a similar example to the length of the empty list can be given at the type level:

type family Length (l :: [k]) :: Nat type instance Length [] = Zero

When Length is applied to the empty (promoted) list it will have the kind Length AnyK [].

AnyK is currently not exported and cannot be used directly, but you might see it in debug output from the compiler.

unsafeCoerce# :: a -> b

The function unsafeCoerce# allows you to side-step the typechecker entirely. That is, it allows you to coerce any type into any other type. If you use this function, you had better get it right, otherwise segmentation faults await. It is generally used when you want to write a program that you know is well-typed, but where Haskell's type system is not expressive enough to prove that it is well typed.

The following uses of unsafeCoerce# are supposed to work (i.e. not lead to spurious compile-time or run-time crashes):

Casting any lifted type to Any

Casting Any back to the real type

Casting an unboxed type to another unboxed type of the same size (but not coercions between floating-point and integral types)

Casting between two types that have the same runtime representation. One case is when the two types differ only in "phantom" type parameters, for example Ptr Int to Ptr Float, or [Int] to [Float] when the list is known to be empty. Also, a newtype of a type T has the same representation at runtime as T.

Other uses of unsafeCoerce# are undefined. In particular, you should not use unsafeCoerce# to cast a T to an algebraic data type D, unless T is also an algebraic data type. For example, do not cast Int->Int to Bool, even if you later cast that Bool back to Int->Int before applying it. The reasons have to do with GHC's internal representation details (for the congnoscenti, data values can be entered but function closures cannot). If you want a safe type to cast things to, use Any, which is not an algebraic data type.

traceEvent# :: Addr# -> State# s -> State# s

Emits an event via the RTS tracing framework. The contents of the event is the zero-terminated byte string passed as the first argument. The event will be emitted either to the .eventlog file, or to stderr, depending on the runtime RTS flags.