GHC.Tuple

Description

The tuple data types

Users should not import this module. It is GHC internal only.

Documentation

data Solo a

Solo is the canonical lifted 1-tuple, just like (,) is the canonical lifted 2-tuple (pair) and (,,) is the canonical lifted 3-tuple (triple).

The most important feature of Solo is that it is possible to force its "outside" (usually by pattern matching) without forcing its "inside", because it is defined as a datatype rather than a newtype. One situation where this can be useful is when writing a function to extract a value from a data structure. Suppose you write an implementation of arrays and offer only this function to index into them:

index :: Array a -> Int -> a

Now imagine that someone wants to extract a value from an array and store it in a lazy-valued finite map/dictionary:

insert "hello" (arr index 12) m

This can actually lead to a space leak. The value is not actually extracted from the array until that value (now buried in a map) is forced. That means the entire array may be kept live by just that value! Often, the solution is to use a strict map, or to force the value before storing it, but for some purposes that's undesirable.

One common solution is to include an indexing function that can produce its result in an arbitrary Applicative context:

indexA :: Applicative f => Array a -> Int -> f a

When using indexA in a pure context, Solo serves as a handy Applicative functor to hold the result. You could write a non-leaky version of the above example thus:

case arr indexA 12 of
  Solo a -> insert "hello" a m

While such simple extraction functions are the most common uses for unary tuples, they can also be useful for fine-grained control of strict-spined data structure traversals, and for unifying the implementations of lazy and strict mapping functions.

data Unit

The unit datatype Unit has one non-undefined member, the nullary constructor ().

Since: ghc-prim-0.11.0

getSolo :: Solo a -> a

Extract the value from a Solo. Very often, values should be extracted directly using pattern matching, to control just what gets evaluated when. getSolo is for convenience in situations where that is not the case:

When the result is passed to a strict function, it makes no difference whether the pattern matching is done on the "outside" or on the "inside":

Data.Set.insert (getSolo sol) set === case sol of Solo v -> Data.Set.insert v set

A traversal may be performed in Solo in order to control evaluation internally, while using getSolo to extract the final result. A strict mapping function, for example, could be defined

map' :: Traversable t => (a -> b) -> t a -> t b
map' f = getSolo . traverse ((Solo $!) . f)

type Tuple0 = ()

A tuple of zero elements, a synonym for ().

Since: ghc-prim-0.11.0

type Tuple1 = Solo

A tuple of one element, a synonym for Solo.

Since: ghc-prim-0.11.0

data Tuple2 a b

A tuple of two elements.

Since: ghc-prim-0.11.0

data Tuple3 a b c

A tuple of three elements.

Since: ghc-prim-0.11.0

data Tuple4 a b c d

A tuple of four elements.

Since: ghc-prim-0.11.0

data Tuple5 a b c d e

A tuple of five elements.

Since: ghc-prim-0.11.0

data Tuple6 a b c d e f

A tuple of six elements.

Since: ghc-prim-0.11.0

data Tuple7 a b c d e f g

A tuple of seven elements.

Since: ghc-prim-0.11.0

data Tuple8 a b c d e f g h

A tuple of eight elements.

Since: ghc-prim-0.11.0

data Tuple9 a b c d e f g h i

A tuple of nine elements.

Since: ghc-prim-0.11.0

data Tuple10 a b c d e f g h i j

A tuple of ten elements.

Since: ghc-prim-0.11.0

data Tuple11 a b c d e f g h i j k

A tuple of eleven elements.

Since: ghc-prim-0.11.0

data Tuple12 a b c d e f g h i j k l

A tuple of twelve elements.

Since: ghc-prim-0.11.0

data Tuple13 a b c d e f g h i j k l m

A tuple of 13 elements.

Since: ghc-prim-0.11.0

data Tuple14 a b c d e f g h i j k l m n

A tuple of 14 elements.

Since: ghc-prim-0.11.0

data Tuple15 a b c d e f g h i j k l m n o

A tuple of 15 elements.

Since: ghc-prim-0.11.0

data Tuple16 a b c d e f g h i j k l m n o p

A tuple of 16 elements.

Since: ghc-prim-0.11.0

data Tuple17 a b c d e f g h i j k l m n o p q

A tuple of 17 elements.

Since: ghc-prim-0.11.0

data Tuple18 a b c d e f g h i j k l m n o p q r

A tuple of 18 elements.

Since: ghc-prim-0.11.0

data Tuple19 a b c d e f g h i j k l m n o p q r s

A tuple of 19 elements.

Since: ghc-prim-0.11.0

data Tuple20 a b c d e f g h i j k l m n o p q r s t

A tuple of 20 elements.

Since: ghc-prim-0.11.0

data Tuple21 a b c d e f g h i j k l m n o p q r s t u

A tuple of 21 elements.

data Tuple22 a b c d e f g h i j k l m n o p q r s t u v

A tuple of 22 elements.

Since: ghc-prim-0.11.0

data Tuple23 a b c d e f g h i j k l m n o p q r s t u v w

A tuple of 23 elements.

Since: ghc-prim-0.11.0

data Tuple24 a b c d e f g h i j k l m n o p q r s t u v w x

A tuple of 24 elements.

Since: ghc-prim-0.11.0

data Tuple25 a b c d e f g h i j k l m n o p q r s t u v w x y

A tuple of 25 elements.

Since: ghc-prim-0.11.0

data Tuple26 a b c d e f g h i j k l m n o p q r s t u v w x y z

A tuple of 26 elements.

Since: ghc-prim-0.11.0

data Tuple27 a b c d e f g h i j k l m n o p q r s t u v w x y z t26

A tuple of 27 elements.

Since: ghc-prim-0.11.0

data Tuple28 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27

A tuple of 28 elements.

Since: ghc-prim-0.11.0

data Tuple29 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28

A tuple of 29 elements.

Since: ghc-prim-0.11.0

data Tuple30 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29

A tuple of 30 elements.

Since: ghc-prim-0.11.0

data Tuple31 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30

A tuple of 31 elements.

Since: ghc-prim-0.11.0

data Tuple32 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31

A tuple of 32 elements.

Since: ghc-prim-0.11.0

data Tuple33 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32

A tuple of 33 elements.

Since: ghc-prim-0.11.0

data Tuple34 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33

A tuple of 34 elements.

Since: ghc-prim-0.11.0

data Tuple35 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34

A tuple of 35 elements.

Since: ghc-prim-0.11.0

data Tuple36 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35

A tuple of 36 elements.

Since: ghc-prim-0.11.0

data Tuple37 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36

A tuple of 37 elements.

Since: ghc-prim-0.11.0

data Tuple38 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37

A tuple of 38 elements.

Since: ghc-prim-0.11.0

data Tuple39 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38

A tuple of 39 elements.

Since: ghc-prim-0.11.0

data Tuple40 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39

A tuple of 40 elements.

Since: ghc-prim-0.11.0

data Tuple41 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40

A tuple of 41 elements.

Since: ghc-prim-0.11.0

data Tuple42 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41

A tuple of 42 elements.

Since: ghc-prim-0.11.0

data Tuple43 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42

A tuple of 43 elements.

Since: ghc-prim-0.11.0

data Tuple44 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43

A tuple of 44 elements.

Since: ghc-prim-0.11.0

data Tuple45 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44

A tuple of 45 elements.

Since: ghc-prim-0.11.0

data Tuple46 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45

A tuple of 46 elements.

Since: ghc-prim-0.11.0

data Tuple47 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45 t46

A tuple of 47 elements.

Since: ghc-prim-0.11.0

data Tuple48 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45 t46 t47

A tuple of 48 elements.

Since: ghc-prim-0.11.0

data Tuple49 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45 t46 t47 t48

A tuple of 49 elements.

Since: ghc-prim-0.11.0

data Tuple50 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45 t46 t47 t48 t49

A tuple of 50 elements.

Since: ghc-prim-0.11.0

data Tuple51 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45 t46 t47 t48 t49 t50

A tuple of 51 elements.

Since: ghc-prim-0.11.0

data Tuple52 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45 t46 t47 t48 t49 t50 t51

A tuple of 52 elements.

Since: ghc-prim-0.11.0

data Tuple53 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45 t46 t47 t48 t49 t50 t51 t52

A tuple of 53 elements.

Since: ghc-prim-0.11.0

data Tuple54 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45 t46 t47 t48 t49 t50 t51 t52 t53

A tuple of 54 elements.

Since: ghc-prim-0.11.0

data Tuple55 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45 t46 t47 t48 t49 t50 t51 t52 t53 t54

A tuple of 55 elements.

Since: ghc-prim-0.11.0

data Tuple56 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45 t46 t47 t48 t49 t50 t51 t52 t53 t54 t55

A tuple of 56 elements.

Since: ghc-prim-0.11.0

data Tuple57 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45 t46 t47 t48 t49 t50 t51 t52 t53 t54 t55 t56

A tuple of 57 elements.

Since: ghc-prim-0.11.0

data Tuple58 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45 t46 t47 t48 t49 t50 t51 t52 t53 t54 t55 t56 t57

A tuple of 58 elements.

Since: ghc-prim-0.11.0

data Tuple59 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45 t46 t47 t48 t49 t50 t51 t52 t53 t54 t55 t56 t57 t58

A tuple of 59 elements.

Since: ghc-prim-0.11.0

data Tuple60 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45 t46 t47 t48 t49 t50 t51 t52 t53 t54 t55 t56 t57 t58 t59

A tuple of 60 elements.

Since: ghc-prim-0.11.0

data Tuple61 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45 t46 t47 t48 t49 t50 t51 t52 t53 t54 t55 t56 t57 t58 t59 t60

A tuple of 61 elements.

Since: ghc-prim-0.11.0

data Tuple62 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45 t46 t47 t48 t49 t50 t51 t52 t53 t54 t55 t56 t57 t58 t59 t60 t61

A tuple of 62 elements.

Since: ghc-prim-0.11.0

data Tuple63 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45 t46 t47 t48 t49 t50 t51 t52 t53 t54 t55 t56 t57 t58 t59 t60 t61 t62

A tuple of 63 elements.

Since: ghc-prim-0.11.0

data Tuple64 a b c d e f g h i j k l m n o p q r s t u v w x y z t26 t27 t28 t29 t30 t31 t32 t33 t34 t35 t36 t37 t38 t39 t40 t41 t42 t43 t44 t45 t46 t47 t48 t49 t50 t51 t52 t53 t54 t55 t56 t57 t58 t59 t60 t61 t62 t63

A tuple of 64 elements.

Since: ghc-prim-0.11.0

data Solo a

Solo is the canonical lifted 1-tuple, just like (,) is the canonical lifted 2-tuple (pair) and (,,) is the canonical lifted 3-tuple (triple).

The most important feature of Solo is that it is possible to force its "outside" (usually by pattern matching) without forcing its "inside", because it is defined as a datatype rather than a newtype. One situation where this can be useful is when writing a function to extract a value from a data structure. Suppose you write an implementation of arrays and offer only this function to index into them:

index :: Array a -> Int -> a

Now imagine that someone wants to extract a value from an array and store it in a lazy-valued finite map/dictionary:

insert "hello" (arr index 12) m

This can actually lead to a space leak. The value is not actually extracted from the array until that value (now buried in a map) is forced. That means the entire array may be kept live by just that value! Often, the solution is to use a strict map, or to force the value before storing it, but for some purposes that's undesirable.

One common solution is to include an indexing function that can produce its result in an arbitrary Applicative context:

indexA :: Applicative f => Array a -> Int -> f a

When using indexA in a pure context, Solo serves as a handy Applicative functor to hold the result. You could write a non-leaky version of the above example thus:

case arr indexA 12 of
  Solo a -> insert "hello" a m

While such simple extraction functions are the most common uses for unary tuples, they can also be useful for fine-grained control of strict-spined data structure traversals, and for unifying the implementations of lazy and strict mapping functions.

Constructors

MkSolo a

Bundled Patterns

pattern Solo :: a -> Solo a

Deprecated: The Solo constructor has been renamed to MkSolo to avoid punning.