GHC.Classes

Description

Basic classes. Do not import this module directly. It is an GHC internal only module. Some of its contents are instead available from Prelude and GHC.Int.

Implicit paramaters

class IP (x :: Symbol) a | x -> a where

The syntax ?x :: a is desugared into IP "x" a IP is declared very early, so that libraries can take advantage of the implicit-call-stack feature

Methods

ip :: a

Equality and ordering

Do not import these classes from this module. Import them from Prelude instead.

class Eq a where

The Eq class defines equality (==) and inequality (/=). All the basic datatypes exported by the Prelude are instances of Eq, and Eq may be derived for any datatype whose constituents are also instances of Eq.

The Haskell Report defines no laws for Eq. However, instances are encouraged to follow these properties:

Reflexivity

x == x = True

Symmetry

x == y = y == x

Transitivity

if x == y && y == z = True, then x == z = True

Extensionality

if x == y = True and f is a function whose return type is an instance of Eq, then f x == f y = True

Negation

x /= y = not (x == y)

Minimal complete definition: either == or /=.

Minimal complete definition

(==) | (/=)

Methods

(==) :: a -> a -> Bool infix 4

(/=) :: a -> a -> Bool infix 4

class Eq a => Ord a where

The Ord class is used for totally ordered datatypes.

Instances of Ord can be derived for any user-defined datatype whose constituent types are in Ord. The declared order of the constructors in the data declaration determines the ordering in derived Ord instances. The Ordering datatype allows a single comparison to determine the precise ordering of two objects.

Ord, as defined by the Haskell report, implements a total order and has the following properties:

Comparability

x <= y || y <= x = True

Transitivity

if x <= y && y <= z = True, then x <= z = True

Reflexivity

x <= x = True

Antisymmetry

if x <= y && y <= x = True, then x == y = True

The following operator interactions are expected to hold:

x >= y = y <= x

x < y = x <= y && x /= y

x > y = y < x

x < y = compare x y == LT

x > y = compare x y == GT

x == y = compare x y == EQ

min x y == if x <= y then x else y = True

max x y == if x >= y then x else y = True

Note that (7.) and (8.) do not require min and max to return either of their arguments. The result is merely required to equal one of the arguments in terms of (==).

Minimal complete definition: either compare or <=. Using compare can be more efficient for complex types.

Minimal complete definition

compare | (<=)

Methods

compare :: a -> a -> Ordering

(<) :: a -> a -> Bool infix 4

(<=) :: a -> a -> Bool infix 4

(>) :: a -> a -> Bool infix 4

(>=) :: a -> a -> Bool infix 4

max :: a -> a -> a

min :: a -> a -> a

Monomorphic equality operators

Matching on class methods (e.g. (==)) in rewrite rules tends to be a bit fragile. For instance, consider this motivating example from the bytestring library,

break :: (Word8 -> Bool) -> ByteString -> (ByteString, ByteString)
breakByte :: Word8 -> ByteString -> (ByteString, ByteString)
{-# RULES "break -> breakByte" forall a. break (== x) = breakByte x #-}

Here we have two functions, with breakByte providing an optimized implementation of break where the predicate is merely testing for equality with a known Word8. As written, however, this rule will be quite fragile as the (==) class operation rule may rewrite the predicate before our break rule has a chance to fire.

For this reason, most of the primitive types in base have Eq and Ord instances defined in terms of helper functions with inlinings delayed to phase 1. For instance, Word8's Eq instance looks like,

instance Eq Word8 where
    (==) = eqWord8
    (/=) = neWord8

eqWord8, neWord8 :: Word8 -> Word8 -> Bool
eqWord8 (W8# x) (W8# y) = ...
neWord8 (W8# x) (W8# y) = ...
{-# INLINE [1] eqWord8 #-}
{-# INLINE [1] neWord8 #-}

This allows us to save our break rule above by rewriting it to instead match against eqWord8,

{-# RULES "break -> breakByte" forall a. break (eqWord8 x) = breakByte x #-}

Currently this is only done for (==), (/=), (<), (<=), (>), and (>=) for the types in GHC.Word and GHC.Int.

eqInt :: Int -> Int -> Bool

neInt :: Int -> Int -> Bool

eqWord :: Word -> Word -> Bool

neWord :: Word -> Word -> Bool

eqChar :: Char -> Char -> Bool

neChar :: Char -> Char -> Bool

eqFloat :: Float -> Float -> Bool

eqDouble :: Double -> Double -> Bool

Monomorphic comparison operators

gtInt :: Int -> Int -> Bool

geInt :: Int -> Int -> Bool

leInt :: Int -> Int -> Bool

ltInt :: Int -> Int -> Bool

compareInt :: Int -> Int -> Ordering

compareInt# :: Int# -> Int# -> Ordering

gtWord :: Word -> Word -> Bool

geWord :: Word -> Word -> Bool

leWord :: Word -> Word -> Bool

ltWord :: Word -> Word -> Bool

compareWord :: Word -> Word -> Ordering

compareWord# :: Word# -> Word# -> Ordering

Functions over Bool

Do not import these functions from this module. Import them from Prelude instead.

(&&) :: Bool -> Bool -> Bool infixr 3

Boolean "and", lazy in the second argument

(||) :: Bool -> Bool -> Bool infixr 2

Boolean "or", lazy in the second argument

not :: Bool -> Bool

Boolean "not"

Integer arithmetic

divInt# :: Int# -> Int# -> Int#

divInt8# :: Int8# -> Int8# -> Int8#

divInt16# :: Int16# -> Int16# -> Int16#

divInt32# :: Int32# -> Int32# -> Int32#

modInt# :: Int# -> Int# -> Int#

modInt8# :: Int8# -> Int8# -> Int8#

modInt16# :: Int16# -> Int16# -> Int16#

modInt32# :: Int32# -> Int32# -> Int32#

divModInt# :: Int# -> Int# -> (# Int#, Int# #)

divModInt8# :: Int8# -> Int8# -> (# Int8#, Int8# #)

divModInt16# :: Int16# -> Int16# -> (# Int16#, Int16# #)

divModInt32# :: Int32# -> Int32# -> (# Int32#, Int32# #)