GHC.Magic

Description

GHC magic.

Use GHC.Exts from the base package instead of importing this module directly.

Documentation

inline :: a -> a

The call '(inline f)' reduces to f, but inline has a BuiltInRule that tries to inline f (if it has an unfolding) unconditionally The NOINLINE pragma arranges that inline only gets inlined (and hence eliminated) late in compilation, after the rule has had a good chance to fire.