GHC.Prim.Ext

Description

Extra C-- routines exposed from the RTS

Actual primops are emitted by the compiler itself. They are special bits of code with backend support. The foreign functions in this module aren't actual primops because the compiler doesn't care about them at all: they just are extra foreign C-- calls libraries can make into the RTS.

Note that Prim has the same haddock section names as this module, but with descriptions. Consult that module's documentation for what each section means. are described over there.

Documentation

type INT64 = Int#

type WORD64 = Word#

Delay/wait operations

Misc

getThreadAllocationCounter# :: State# RealWorld -> (# State# RealWorld, INT64 #)

Retrieves the allocation counter for the current thread.